+++ rust/ballista/src/physical_plan.rs
@@ -12,19 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Ballista Physical Plan (Experimental)
+//! Ballista Physical Plan (Experimental).
+//!
+//! The physical plan is a serializable data structure describing how the plan will be executed.
+//!
+//! It differs from the logical plan in that it deals with specific implementations of operators
+//! (e.g. SortMergeJoin versus BroadcastHashJoin) whereas the logical plan just deals with an
+//! abstract concept of a join.
+//!
+//! The physical plan also accounts for partitioning and ordering of data between operators.
 
 use crate::arrow::datatypes::Schema;
 use crate::arrow::record_batch::RecordBatch;
 use crate::datafusion::logicalplan::LogicalPlan;
-
-///
-#[derive(Debug, Clone)]
-pub enum Action {
-    Collect { plan: LogicalPlan },
-    WriteCsv { plan: LogicalPlan, path: String },
-    WriteParquet { plan: LogicalPlan, path: String },
-}
 
 #[derive(Debug, Clone)]
 pub enum PhysicalPlan {
@@ -149,50 +149,5 @@ pub struct HashAggregatePlan {
 #[derive(Debug, Clone)]
 pub enum Expression {
     Column(usize),
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////////////
-// Execution (interpreted) below this point
-///////////////////////////////////////////////////////////////////////////////////////////////////
-
-pub struct ColumnarBatch {
-    /// we just wrap Arrow RecordBatch for now, but this may change later so that we can
-    /// support scalar values as well as columnar results
-    record_batch: RecordBatch,
-}
-
-pub trait ExecutionPlan {
-    /// Specifies how data is partitioned across different nodes in the cluster
-    fn output_partitioning(&self) -> Partitioning;
-    /// Specifies how data is ordered in each partition
-    fn output_ordering(&self) -> Vec<SortOrder>;
-    /// Specifies the data distribution requirements of all the children for this operator
-    fn required_child_ordering(&self) -> Vec<Vec<SortOrder>>;
-    /// Runs this query returning a stream of colymnar batches
-    fn execute(&self); //TODO decide on return type to represent stream of record batches
-    /// Runs this query returning the full results
-    fn execute_collect(&self) -> Vec<ColumnarBatch>;
-    /// Runs this query returning the first `n` rows
-    fn execute_take(&self, n: usize) -> Vec<ColumnarBatch>;
-    /// Runs this query returning the last `n` rows
-    fn execute_tail(&self, n: usize) -> Vec<ColumnarBatch>;
-    /// Returns the children of this operator
-    fn children(&self) -> Vec<Box<dyn ExecutionPlan>>;
-}
-
-pub trait UnaryExec: ExecutionPlan {
-    fn child(&self) -> Box<dyn ExecutionPlan>;
-
-    fn children(&self) -> Vec<Box<dyn ExecutionPlan>> {
-        vec![self.child()]
-    }
-}
-
-pub trait BinaryExec: ExecutionPlan {
-    fn left(&self) -> Box<dyn ExecutionPlan>;
-    fn right(&self) -> Box<dyn ExecutionPlan>;
-
-    fn children(&self) -> Vec<Box<dyn ExecutionPlan>> {
-        vec![self.left(), self.right()]
-    }
+    //TODO: add all the expressions here or possibly re-use the expressions from the logical plan
 }
